Fix playerChoice validation and create round logic
Function getPlayerChoice used prompt to get user choice and if choice
matched any value in VALID_CHOICES it returned player choice or alert
message. VALID_CHOICES values are in lower case so player choice was
converted to lower case.

The problem is that player choice is not always a string.
Also after adding a function that returns round result there was a
problem where this function was returning value even tho player choice
was not valid value.

Now function isPlayerChoiceValid checks if type of player choice is
string so there is no more problem with player choice to lower case.
This function also solves second problem because now we run
getRoundResult only if player choice is a valid value.