Fix user input validation and new functionality
Changes in completeGameRound function:

isPlayerChoiceValid function was checking type of user input and if
input was in valid pool and it decided if playerChoice was a valid value
toLowerCase conversion on playChoice was failing because playerChoice
could have different type than string. Now this function checks type
before allowing toLowerCase conversion and returns false if its
different type than string.

added if that checks if cancel button was pressed which allows us to
exit program

playerChoice and pcChoice variables are moved to startGame function so
we can access them and display them to user.

Added functionality:

player chooses how many rounds he wants to play
validation of how many rounds choice

startGame is a main function, which main task is to get result of the
round and act accordingly to received value