Add sidecontent and create exit to main menu func
Added container around main content and side content which allows to
display side content to the side of main content.

In js file created new function mainMenuExit which is invoked when user
clicks exit button in top right corner. It's task is to add disappear
class to every item and remove disappear class from main menu items.
To do that it is using showOrHide function.
The problem was that showOrHide function is toggling disappear class, so
images that had disappeared would actually appear in main menu.
To solve that showOrHide function had to take callback function instead
of element to filter out, which allowed to filter out images that
already had disappear class.